backend/requirements.txt
Django
psycopg[binary]
argon2-cffi
zxcvbn
djangorestframework
djangorestframework-simplejwt
drf-spectacular
pillow
python-dotenv
black 
isort
flake8
flake8-bugbear
flake8-bandit
mccabe
pytest
pytest-django
coverage
django-cors-headers